requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyJWT
pydantic[dotenv]
httpx
openai
azure-cosmos
aiohttp
cryptography
